Spatial-text search
A program that will read the data and create the inverted file and grid implemented in Parts 1 and 2.

kwSpaSearchGrid function takes as a argument a query range and a list of query keywords and calculates and returns the restaurants contained
in the query range and contain all the query keywords in their tags. For that reason it uses the grid to find the restaurants contained in the query range and for
each of them verifies if it contains all the query keywords in its tags.

Note: A divide and conquer style was implemented to speed up the merge-join function